Corrected sound frequencies
Playing a 440 Hz sound (by writing the value $4A to the YM2151 register $28) did not produce a sound output of the desired frequency, but rather 246 Hz. This was verified using both audacity as well as an external guitar tuner.

Frequencies generated by the YM2151 are determined based on the chip clock frequency, which according to the YM2151 documentation is expected to be 3.579545 MHz.

For some (for me unknown) reason the value needed in line 468 is apparently twice the chip clock frequency. With this single change, the sound output now has the desired frequency, verified in both audacity and an external guitar tuner.