Fix LOAD/SAVE hypercall so debugger doesn't break
Attempting to step-over or step-into a `jsr $ffd5` resulted in the debugger not having the opportunity to pause execution until after whatever instruction followed the `jsr`.  In the case of the step-over, since the debugger looks for a specific address to pause execution and that address was being processed before the debugger could pause execution, execution would simply continue without the debugger pausing appropriately.

By inserting a "continue" at the end of the hypercall's if-block, the debugger gets it chance to pause execution properly.