Trigger VSYNC as soon as scan goes offscreen
The VSYNC interrupt on the Commander X16 should trigger as soon as the
current scanline is offscreen. Currently, the interrupt only triggers
once the scan returns to the top of the screen, wasting many lines that
could otherwise be spent performing time-sensitive operations. This
change to video.c causes the interrupt to trigger at the proper
location.