Change end of screen check to use y
When checking whether the end of the screen has been reached, the
variable y should be used in place of scan_pos_y. This is because y does
not include the front porch region, which, if included, would cause the
end of the screen to trigger several lines too early.