Fix VSYNC timing (#322)
* Trigger VSYNC as soon as scan goes offscreen

The VSYNC interrupt on the Commander X16 should trigger as soon as the
current scanline is offscreen. Currently, the interrupt only triggers
once the scan returns to the top of the screen, wasting many lines that
could otherwise be spent performing time-sensitive operations. This
change to video.c causes the interrupt to trigger at the proper
location.

* Change end of screen check to use y

When checking whether the end of the screen has been reached, the
variable y should be used in place of scan_pos_y. This is because y does
not include the front porch region, which, if included, would cause the
end of the screen to trigger several lines too early.

* Move incrementation of scan_pos_y