Fix issue #353 (#354)
Cap hstart and hstop so that the emulator will no longer attempt to write data outside of the line buffer.

Also removing the "LAYER_PIXELS_PER_ITERATION" stuff, that nobody is convinced is a win, and I'm pretty sure this simplified version is a little faster, even, with a slight tweak to how it iterates the scaled "eff_x" index into the display buffer. Certainly, this provably saves us some work by not doing a bunch of drawing work and then clobbering pixels with a border after the fact.