Added testbench: running emulator headless for unit testing (#405)
This adds a new switch to the emulator ("-testbench").

When enabled, the emulator runs in headless mode. The feature is made for unit testing by a test runner typically made in a high level language such as python.

During testbench mode, the emulator listens on stdin for commands sent by a test runner. Those commands may be used to set the state of the emulator (registers, memory address values) before you run the code that is to be tested.

After the test is completed, the test runner may check the state of the emulator (registers, memory address values). The emulator sends its responses over stdout.

The communication "protocol" between the emulator and the test runner is described at the top of the file testbench.h.

There is also a python example implementing a unit test:

The file testbench.py defines a class ("X16TestBench") that makes communicating with the emulator a bit easier
The file test.py is a unit test, checking the result of an addition
This code is quite new, and I'm sure there are still bugs to be found.